requirements.txt
fastapi
uvicorn
aiogram
supabase
python-dotenv
pydantic-settings
groq
apscheduler
cryptography
google-auth
google-auth-oauthlib
google-api-python-client
